Remove mention of distro for non-cloud playbook
These example playbooks are distro-independent, because the requirements state that
there exist properly configured machines on to which the add-on installation should
be done. Examples in the actual collection repository additionally include prepare
scripts to bootstrap an infrastructure to easily test the Ansible add-on installer.
But this falls outside the scope of purposes for this repository.